% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ach}
\alias{ach}
\title{Achilles example data set}
\format{
An object of class \code{Achilles} with 100 rows and 50 columns.
}
\usage{
ach
}
\description{
Achilles example data set
}
\note{
Updated 2020-10-07.
}
\examples{
data(ach)
ach
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
