% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Achilles-class}
\alias{Achilles-class}
\title{Achilles}
\value{
\code{Achilles}.
}
\description{
Legacy object that is no longer recommended.
}
\details{
Use \code{GeneEffect} object instead.
}
\note{
Updated 2021-07-07.
}
\keyword{internal}
