% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles.R
\name{Achilles}
\alias{Achilles}
\title{Import Project Achilles CRISPR gene effect data}
\usage{
Achilles(release = NULL, rowRanges = TRUE, colData = TRUE)
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}

\item{rowRanges}{\code{logical(1)}.
Include gene annotations as genomic ranges.}

\item{colData}{\code{logical(1)}.
Include cell line metadata.}
}
\value{
\code{Achilles}.
}
\description{
Import Project Achilles CRISPR gene effect data
}
\note{
Updated 2020-10-01.
}
\section{Assays}{

\itemize{
\item \code{effect}: \strong{CERES data} with principle components strongly related to
known batch effects removed, then shifted and scaled per cell line so the
median nonessential KO effect is 0 and the median essential KO effect is
-1.
\item \code{probability}: \strong{Probability} that knocking out the gene has a real
depletion effect using \code{gene_effect}.
}
}

\examples{
object <- Achilles()
print(object)
}
