% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AchillesGeneDependencyData-class}
\alias{AchillesGeneDependencyData-class}
\title{Achilles gene dependency data}
\value{
\code{AchillesGeneDependencyData}.
}
\description{
Achilles gene dependency data
}
\note{
Updated 2020-09-30.
}
