% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AchillesGeneEffectData-class}
\alias{AchillesGeneEffectData-class}
\title{Achilles gene effect data}
\value{
\code{AchillesGeneEffectData}.
}
\description{
Achilles gene effect data
}
\note{
Updated 2020-09-30.
}
