% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{AchillesGeneEffectData}
\alias{AchillesGeneEffectData}
\title{Import Achilles gene effect data}
\usage{
AchillesGeneEffectData(release = NULL)
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}
}
\value{
\code{AchillesGeneEffectData}.
}
\description{
Import Achilles gene effect data
}
\note{
Updated 2020-09-30.
}
\examples{
object <- AchillesGeneEffectData()
dim(object)
}
