% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CCLECopyNumberData-class}
\alias{CCLECopyNumberData-class}
\title{CCLE copy number data}
\value{
\code{CCLECopyNumberData}.
}
\description{
CCLE copy number data
}
\note{
Updated 2021-02-25.
}
