% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CCLEExpressionData-class}
\alias{CCLEExpressionData-class}
\title{CCLE expression data}
\value{
\code{CCLEExpressionData}.
}
\description{
CCLE expression data
}
\note{
Updated 2020-09-30.
}
