% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEExpressionData.R
\name{CCLEExpressionData}
\alias{CCLEExpressionData}
\title{Import CCLE expression data}
\usage{
CCLEExpressionData(dataset = "depmap_public_22q2")
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{CCLEExpressionData}.
}
\description{
Import CCLE expression data
}
\note{
Updated 2021-07-07.
}
\examples{
object <- CCLEExpressionData()
dim(object)
}
