% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEFusionData.R
\name{CCLEFusionData}
\alias{CCLEFusionData}
\title{Import CCLE fusion data}
\usage{
CCLEFusionData(dataset = "depmap_public_22q2", filtered = TRUE)
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}

\item{filtered}{\code{logical(1)}.
Load filtered fusion calls.}
}
\value{
\code{CCLEFusionData}.
}
\description{
Import CCLE fusion data
}
\note{
Updated 2022-03-09.
}
\examples{
object <- CCLEFusionData()
dim(object)
}
