% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CCLEMicroRNAExpressionData-class}
\alias{CCLEMicroRNAExpressionData-class}
\title{CCLE microRNA expression data}
\value{
\code{CCLEMicroRNAExpressionData}.
}
\description{
CCLE microRNA expression data
}
\details{
Inherits from \code{SummarizedExperiment}.
Cells in columns, microRNAs in rows.

Same as the re-released \code{CCLE_miRNA_20181103.gct} file.
}
\note{
Updated 2023-01-25.
}
