% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEMicroRNAExpressionData.R
\name{CCLEMicroRNAExpressionData}
\alias{CCLEMicroRNAExpressionData}
\title{Import CCLE microRNA expression data}
\usage{
CCLEMicroRNAExpressionData()
}
\value{
\code{CCLEMicroRNAExpressionData}.
}
\description{
Import CCLE microRNA expression data
}
\details{
NanoString microRNA panel data.
}
\note{
Updated 2022-11-16.
}
\examples{
object <- CCLEMicroRNAExpressionData()
dim(object)
}
