% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CCLEMutationData-class}
\alias{CCLEMutationData-class}
\title{CCLE mutation data}
\value{
\code{CCLEMutationData}.
}
\description{
CCLE mutation data
}
\details{
Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2021-02-25.
}
