% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE.R
\name{CCLEMutationData}
\alias{CCLEMutationData}
\title{Import CCLE mutation data}
\usage{
CCLEMutationData(release = "21Q1")
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}
}
\value{
\code{CCLEMutationData}.
}
\description{
Import CCLE mutation data
}
\note{
Updated 2021-02-24.
}
\examples{
object <- CCLEMutationData()
dim(object)
}
