% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEProteomicsData.R
\name{CCLEProteomicsData}
\alias{CCLEProteomicsData}
\title{CCLE proteomics data}
\usage{
CCLEProteomicsData()
}
\description{
CCLE proteomics data
}
\note{
Updated 2022-09-21.
}
\seealso{
\itemize{
\item https://depmap.org/portal/download/all/?releasename=Proteomics
\item https://gygi.hms.harvard.edu/publications/ccle.html
}
}
