% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellLineSampleData-class}
\alias{CellLineSampleData-class}
\title{Cell line sample metadata}
\value{
\code{CellLineSampleData}.
}
\description{
Cell line sample metadata
}
\note{
Updated 2020-09-30.
}
