% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{CellLineSampleData}
\alias{CellLineSampleData}
\title{Import cell line sample metadata}
\usage{
CellLineSampleData(release = NULL)
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}
}
\value{
\code{DataFrame}.
}
\description{
Import cell line sample metadata
}
\note{
Updated 2020-09-30.
}
\examples{
object <- CellLineSampleData()
dim(object)
}
