% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Codependencies-class}
\alias{Codependencies-class}
\title{Gene effect co-dependencies}
\value{
\code{Codependencies}.
}
\description{
Gene effect co-dependencies
}
\details{
Inherits from \code{DataFrame}.
}
\note{
Updated 2021-07-19.
}
