% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Codependencies-methods.R
\name{Codependencies}
\alias{Codependencies}
\alias{Codependencies,GeneEffect}
\alias{Codependencies,GeneEffect-method}
\title{Calculate co-dependency scores}
\usage{
Codependencies(object, ...)

`Codependencies,GeneEffect`(object, geneName1, geneName2 = NULL)

\S4method{Codependencies}{GeneEffect}(object, geneName1, geneName2 = NULL)
}
\arguments{
\item{geneName1}{\code{character(1)}.
Gene name.}

\item{geneName2}{\code{character(1)} or \code{NULL}.
Gene name or \code{NULL}.
If \code{NULL}, calculate correlations against all genes.}
}
\value{
\code{Codependencies}.
}
\description{
Calculate co-dependency scores
}
\note{
Updated 2022-08-05.
}
\examples{
data(crispr)

## GeneEffect ====
object <- crispr

## Calculate all co-dependencies for a gene of interest.
x <- Codependencies(
    object = object,
    geneName1 = "EZH2",
    geneName2 = NULL
)
print(head(x))

## Lineage restrict to prostate cancer.
disease <- "Prostate carcinoma"
keep <- colData(object)[["cellosaurusNcItDisease"]] \%in\% disease
object <- object[, keep]
x <- Codependencies(
    object = object,
    geneName1 = "EZH2",
    geneName2 = "EED"
)
print(x)
}
\seealso{
\itemize{
\item \href{https://forum.depmap.org/t/down-load-of-co-dependencies/175/}{How to download codependency scores}
\item \href{https://gist.github.com/pgm/ac2ac4c664ef81200ce49133cc4cee02}{Python script for processing of all codependency scores}
}
}
