% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyNumber.R
\name{CopyNumber}
\alias{CopyNumber}
\title{Import CCLE copy number data}
\usage{
CopyNumber(dataset = "depmap_public_22q2")
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{CopyNumber}.
}
\description{
Import CCLE copy number data
}
\note{
Updated 2023-01-26.
}
\examples{
object <- CopyNumberData()
print(object)
}
