% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crispr}
\alias{crispr}
\title{CRISPR example data set}
\format{
An object of class \code{DepMapCRISPRGeneEffect} of length 1.
}
\usage{
data(crispr)
}
\value{
\code{DepMapGeneEffect}.
}
\description{
CRISPR example data set
}
\note{
Updated 2023-01-26.
}
\examples{
data(crispr)
print(crispr)
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
