% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{DepMap datasets}
\format{
An object of class \code{list} of length 9.
}
\usage{
data(datasets)
}
\value{
\code{list}.
}
\description{
DepMap datasets
}
\note{
Updated 2021-07-15.
}
\examples{
data(datasets)
print(names(datasets))
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
