% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem}
\alias{dem}
\title{DEMETER2 example data set}
\format{
An object of class \code{DEMETER2} of length 1.
}
\usage{
dem
}
\description{
DEMETER2 example data set
}
\note{
Updated 2021-02-24.
}
\examples{
data(dem)
dem
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
