% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DEMETER2-class}
\alias{DEMETER2-class}
\title{DEMETER2 RNAi gene effect data}
\value{
\code{DEMETER2}.
}
\description{
DEMETER2 RNAi gene effect data
}
\details{
Cells in columns, genes in rows.
}
\note{
Updated 2021-02-24.
}
