% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEMETER2.R
\name{DEMETER2}
\alias{DEMETER2}
\title{Import DEMETER2 RNAi screen gene effect data}
\usage{
DEMETER2(rowData = TRUE, colData = TRUE)
}
\value{
\code{DEMETER2}.
}
\description{
Import DEMETER2 RNAi screen gene effect data
}
\note{
Updated 2021-02-25.
}
\examples{
object <- DEMETER2()
dim(object)
}
