% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DEMETER2GeneEffectData-class}
\alias{DEMETER2GeneEffectData-class}
\title{DEMETER2 RNAi gene effect data}
\value{
\code{DEMETER2GeneEffectData}.
}
\description{
DEMETER2 RNAi gene effect data
}
\note{
Updated 2020-09-30.
}
