% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{DEMETER2GeneEffectData}
\alias{DEMETER2GeneEffectData}
\title{Import DEMETER2 RNAi screen gene effect data}
\usage{
DEMETER2GeneEffectData()
}
\value{
\code{DEMETER2GeneEffectData}.
}
\description{
Import DEMETER2 RNAi screen gene effect data
}
\note{
Updated 2020-09-30.
}
\examples{
object <- DEMETER2GeneEffectData()
dim(object)
}
