% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapAnalysis-class}
\alias{DepMapAnalysis-class}
\title{Cancer cell line dependency map analysis}
\value{
\code{CERES}.
}
\description{
Cancer cell line dependency map analysis
}
\details{
Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2021-06-09.
}
\seealso{
\itemize{
\item https://depmap.org/portal/achilles/
\item https://depmap.org/ceres/
}
}
