% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{DepMapAnalysis-package}
\title{DepMapAnalysis}
\description{
Cancer Dependency Map (DepMap) analysis toolkit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/depmapanalysis/}
  \item \url{https://github.com/acidgenomics/r-depmapanalysis/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-depmapanalysis/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Authors:
\itemize{
  \item Teresa Rice \email{teresapillars@gmail.com} (\href{https://orcid.org/0000-0001-8641-6303}{ORCID})
  \item Nitesh Turaga \email{mail@niteshturaga.com} (\href{https://orcid.org/0000-0002-0224-9817}{ORCID})
}

Other contributors:
\itemize{
  \item Redona Therapeutics [funder]
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
