% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapAnalysis.R
\name{DepMapAnalysis}
\alias{DepMapAnalysis}
\title{Import cancer cell line dependency map data}
\usage{
DepMapAnalysis(
  dataset = c("depmap_public_21q2", "depmap_public_21q1", "depmap_public_20q4v2",
    "depmap_public_20q4", "depmap_public_20q3", "demeter2_data_v6"),
  project = c("combined", "achilles"),
  scoringMethod = c("chronos", "ceres", "demeter2"),
  rowData = TRUE,
  colData = TRUE
)
}
\arguments{
\item{dataset}{\code{character(1)} or \code{NULL}.
DepMap dataset release name.
If left \code{NULL}, defaults to latest supported dataset.}

\item{project}{\code{character(1)}.
Project name.
Defaults to a combination of multiple projects (\code{"combined"}):
\itemize{
\item CRISPR: Broad DepMap Public, Sanger Project Score
(e.g. for \code{depmap_public_21q2}).
\item RNAi: Achilles, DRIVE, Marcotte
(e.g. for \code{demeter2_data_v6}).
}}

\item{scoringMethod}{\code{character(1)}.}

\item{rowData}{\code{logical(1)}.
Include gene annotations as genomic ranges.}

\item{colData}{\code{logical(1)}.
Include cell line metadata.}
}
\value{
\code{DepMapAnalysis}.
}
\description{
Import cancer cell line dependency map data
}
\note{
Updated 2021-06-09.
}
\section{Assays}{

\itemize{
\item \code{effect}: \strong{CERES data} with principle components strongly related to
known batch effects removed, then shifted and scaled per cell line so the
median nonessential KO effect is 0 and the median essential KO effect is
-1.
\item \code{probability}: \strong{Probability} that knocking out the gene has a real
depletion effect using \code{gene_effect}.
}
}

\examples{
object <- Achilles()
print(object)
}
