% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapCodependencies-class}
\alias{DepMapCodependencies-class}
\title{DepMap gene effect co-dependencies}
\value{
\code{DepMapCodependencies}.
}
\description{
DepMap gene effect co-dependencies
}
\details{
Inherits from \code{DataFrame}.
}
\note{
Updated 2023-01-26.
}
