% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/DepMapCodependencies-methods.R
\name{DepMapCodependencies}
\alias{DepMapCodependencies}
\alias{DepMapCodependencies,DepMapGeneEffect-method}
\title{Calculate co-dependency scores}
\usage{
DepMapCodependencies(object, ...)

\S4method{DepMapCodependencies}{DepMapGeneEffect}(object, gene1, gene2 = NULL)
}
\arguments{
\item{gene1}{\code{character(1)}.
Gene identifier.}

\item{gene2}{\code{character(1)} or \code{NULL}.
Gene identifier or \code{NULL}.
If \code{NULL}, calculate correlations against all genes.}
}
\value{
\code{DepMapCodependencies}.
}
\description{
Calculate co-dependency scores
}
\note{
Updated 2023-02-03.
}
\examples{
data(crispr)

## DepMapGeneEffect ====
object <- crispr
genes <- as.character(rowData(object)[["geneName"]])
gene1 <- genes[[1L]]
gene2 <- genes[[2L]]
print(c(gene1, gene2))
## Calculate all co-dependencies for a gene of interest.
x <- DepMapCodependencies(
    object = object,
    gene1 = 1,
    gene2 = NULL
)
print(head(x))
## Lineage restrict and compare 2 genes.
diseases <- as.character(colData(object)[["cellosaurusNcItDisease"]])
disease <- diseases[[1L]]
keep <- colData(object)[["cellosaurusNcItDisease"]] \%in\% disease
object2 <- object[, keep]
x <- DepMapCodependencies(
    object = object2,
    gene1 = gene1,
    gene2 = gene2
)
print(x)
}
\seealso{
\itemize{
\item \href{https://forum.depmap.org/t/down-load-of-co-dependencies/175/}{How to download codependency scores}
\item \href{https://gist.github.com/pgm/ac2ac4c664ef81200ce49133cc4cee02}{Python script for processing of all codependency scores}
}
}
