% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapCopyNumber-class}
\alias{DepMapCopyNumber-class}
\title{DepMap (CCLE) copy number data}
\value{
\code{DepMapCopyNumber}.
}
\description{
DepMap (CCLE) copy number data
}
\details{
Gene level copy number data, log2 transformed with a pseudo count of 1.
This is generated by mapping genes onto the segment level calls.

Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2023-01-26.
}
