% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapCRISPRGeneEffect-class}
\alias{DepMapCRISPRGeneEffect-class}
\title{DepMap CRISPR gene effect}
\value{
\code{DepMapCRISPRGeneEffect}.
}
\description{
DepMap CRISPR gene effect
}
\details{
Inherits from \code{DepMapGeneEffect}, which extends \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2023-08-09.
}
\seealso{
\itemize{
\item https://depmap.org/portal/achilles/
\item https://depmap.org/ceres/
\item https://score.depmap.sanger.ac.uk/
}
}
