% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapExperiment-class}
\alias{DepMapExperiment-class}
\title{\code{DepMapExperiment} virtual class}
\value{
\code{DepMapExperiment}.
}
\description{
\code{DepMapExperiment} virtual class
}
\details{
Virtual class, not intended to be worked with directly.

Extends \code{SummarizedExperiment} class, with additional DepMap-specific
validity checks.

Cells in columns, genes in rows.
}
\note{
Updated 2023-09-14.
}
