% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapExpression-class}
\alias{DepMapExpression-class}
\title{\code{DepMapExpression} virtual class}
\value{
\code{DepMapExpression}.
}
\description{
\code{DepMapExpression} virtual class
}
\details{
Virtual class, not intended to be worked with directly.

RNA-seq TPM gene expression data for just protein coding genes using RSEM.
Log2 transformed, using a pseudo-count of 1.

Inherits from \code{DepMapExperiment}.
}
\note{
Updated 2023-08-22.
}
