% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapTxExpression.R
\name{DepMapExpression}
\alias{DepMapExpression}
\alias{DepMapGeneExpression}
\alias{DepMapTxExpression}
\title{Import DepMap gene or transcript expression data}
\usage{
DepMapGeneExpression()

DepMapTxExpression()
}
\value{
\code{DepMapExpression}.
}
\description{
Import DepMap gene or transcript expression data
}
\note{
Updated 2023-09-14.
}
\examples{
## Gene level.
object <- DepMapGeneExpression()
print(object)

## Transcript level (CPU intensive).
## > object <- DepMapTxExpression()
## > print(object)
}
