% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapFusion.R
\name{DepMapFusion}
\alias{DepMapFusion}
\title{Import DepMap fusion call data}
\usage{
DepMapFusion(dataset = "depmap_public_23q2", filtered = TRUE)
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}

\item{filtered}{\code{logical(1)}.
Load filtered fusion calls.}
}
\value{
\code{DepMapFusion}.
}
\description{
Import DepMap fusion call data
}
\note{
Updated 2023-01-26.
}
\examples{
object <- DepMapFusion()
dim(object)
}
