% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapFusions-class}
\alias{DepMapFusions-class}
\title{DepMap fusion call data}
\value{
\code{DepMapFusions}.
}
\description{
DepMap fusion call data
}
\details{
Inherits from \code{DFrame}.
}
\note{
Updated 2023-08-02.
}
