% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapFusions.R
\name{DepMapFusions}
\alias{DepMapFusions}
\title{Import DepMap fusion call data}
\usage{
DepMapFusions(dataset = "depmap_public_23q2", filtered = TRUE)
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}

\item{filtered}{\code{logical(1)}.
Load filtered fusion calls.}
}
\value{
\code{DepMapFusions}.
}
\description{
Import DepMap fusion call data
}
\note{
Updated 2023-08-02.
}
\examples{
object <- DepMapFusions()
dim(object)
}
