% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapGeneEffect-class}
\alias{DepMapGeneEffect-class}
\title{Gene effect in cancer cell lines}
\value{
\code{DepMapGeneEffect}.
}
\description{
Gene effect in cancer cell lines
}
\details{
Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2023-01-26.
}
\seealso{
\itemize{
\item https://depmap.org/portal/achilles/
\item https://depmap.org/ceres/
\item https://score.depmap.sanger.ac.uk/
}
}
