% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapGeneEffect.R
\name{DepMapGeneEffect}
\alias{DepMapGeneEffect}
\alias{DepMapCrisprGeneEffect}
\alias{DepMapRnaiGeneEffect}
\title{DepMap gene effect in cancer cell lines}
\usage{
DepMapCrisprGeneEffect()

DepMapRnaiGeneEffect()
}
\value{
\code{DepMapGeneEffect}.
}
\description{
DepMap gene effect in cancer cell lines
}
\note{
Updated 2023-08-09.
}
\section{Assays}{

\itemize{
\item \code{effect}: \strong{Chronos or CERES data} with principle components strongly
related to known batch effects removed, then shifted and scaled per cell
line so the median nonessential KO effect is 0 and the median essential KO
effect is -1.
\item \code{probability}: \strong{Probability} that knocking out the gene has a real
depletion effect using \code{gene_effect}.
}
}

\examples{
## CRISPR.
object <- DepMapCrisprGeneEffect()
print(object)

## RNAi (DEMETER2).
object <- DepMapRnaiGeneEffect()
print(object)
}
