% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapGeneExpression-class}
\alias{DepMapGeneExpression-class}
\title{DepMap RNA-seq gene expression}
\value{
\code{DepMapGeneExpression}.
}
\description{
DepMap RNA-seq gene expression
}
\details{
Inherits from \code{DepMapExpression}, which extends \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2023-08-09.
}
