% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapGeneExpression.R
\name{DepMapGeneExpression}
\alias{DepMapGeneExpression}
\title{Import DepMap gene expression data}
\usage{
DepMapGeneExpression()
}
\value{
\code{DepMapGeneExpression}.
}
\description{
Import DepMap gene expression data
}
\note{
Updated 2023-09-14.
}
\examples{
## Gene level.
object <- DepMapGeneExpression()
print(object)
}
