% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapMicroRNA.R
\name{DepMapMicroRNA}
\alias{DepMapMicroRNA}
\title{Import DepMap microRNA expression data}
\usage{
DepMapMicroRNA()
}
\value{
\code{DepMapMicroRNA}.
}
\description{
Import DepMap microRNA expression data
}
\details{
NanoString microRNA panel data.
}
\note{
Updated 2023-08-08.
}
\examples{
object <- DepMapMicroRNA()
dim(object)
}
