% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapMutation-class}
\alias{DepMapMutation-class}
\title{DepMap mutation data}
\value{
\code{DepMapMutation}.
}
\description{
DepMap mutation data
}
\details{
Inherits from \code{DFrame}.
}
\note{
Updated 2023-01-26.
}
