% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapMutation.R
\name{DepMapMutation}
\alias{DepMapMutation}
\title{Import DepMap mutation data}
\usage{
DepMapMutation(dataset = "depmap_public_23q2")
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{DepMapMutation}.
}
\description{
Import DepMap mutation data
}
\note{
Updated 2023-02-03.
}
\examples{
object <- DepMapMutation()
dim(object)
}
