% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapMutations-class}
\alias{DepMapMutations-class}
\title{DepMap somatic mutation data}
\value{
\code{DepMapMutations}.
}
\description{
DepMap somatic mutation data
}
\details{
Inherits from \code{DFrame}.
}
\note{
Updated 2023-08-02.
}
