% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapMutations.R
\name{DepMapMutations}
\alias{DepMapMutations}
\title{Import DepMap somatic mutation data}
\usage{
DepMapMutations()
}
\value{
\code{DepMapMutations}.
}
\description{
Import DepMap somatic mutation data
}
\note{
Updated 2023-08-09.
}
\examples{
object <- DepMapMutations()
dim(object)
}
