% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapProteomics-class}
\alias{DepMapProteomics-class}
\title{DepMap protein expression data}
\value{
\code{DepMapProteomics}.
}
\description{
DepMap protein expression data
}
\details{
Inherits from \code{SummarizedExperiment}.
Cells in columns, peptides in rows.
}
\note{
Updated 2023-02-03.
}
