% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapProteomics.R
\name{DepMapProteomics}
\alias{DepMapProteomics}
\title{Import DepMap proteomics data}
\usage{
DepMapProteomics(dataset = c("goncalves_2022", "nusinow_2020"))
}
\value{
\code{DepMapProteomics}.
}
\description{
Import DepMap proteomics data
}
\note{
Updated 2023-08-09.
}
\section{Nusinow et al. 2020 (Gygi lab) dataset}{


Citation:
https://doi.org/10.1016/j.cell.2019.12.023
}

\section{UniProt}{


Note that \code{"sp"} for UniProtKB/Swiss-Prot and \code{"tr"} for UniProtKB/TrEMBL.

See also:
\itemize{
\item https://www.uniprot.org/help/fasta-headers
}
}

\seealso{
\itemize{
\item https://depmap.org/portal/download/all/?releasename=Proteomics
\item https://gygi.hms.harvard.edu/publications/ccle.html
}
}
