% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapTxExpression.R
\name{DepMapTxExpression}
\alias{DepMapTxExpression}
\title{Import DepMap transcript expression data}
\usage{
DepMapTxExpression()
}
\value{
\code{DepMapTxExpression}.
}
\description{
Import DepMap transcript expression data
}
\note{
Updated 2023-09-14.
}
\examples{
## This is CPU intensive.
## > object <- DepMapTxExpression()
## > print(object)
}
