% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffExp-methods.R
\name{diffExp}
\alias{diffExp}
\alias{diffExp,DepMapExpression-method}
\title{Differential expression}
\usage{
diffExp(object, ...)

\S4method{diffExp}{DepMapExpression}(object, sensitiveCells, insensitiveCells)
}
\arguments{
\item{object}{Object.}

\item{insensitiveCells, sensitiveCells}{\code{character}.
Cell line names.}
}
\value{
\code{DFrame}.
}
\description{
Differential expression
}
\note{
Updated 2023-08-22.
}
\examples{
data(rnaseq)

## DepMapExpression ====
object <- rnaseq
sensitiveCells <- colnames(object)[c(1L:5L)]
insensitiveCells <- colnames(object)[c(6L:10L)]
x <- findBiomarkers(
    object = object,
    sensitiveCells = sensitiveCells,
    insensitiveCells = insensitiveCells
)
print(x)
}
\seealso{
\itemize{
\item limma User's Guide, Section 15.4.
}
}
