% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/euclidean-methods.R
\name{euclidean}
\alias{euclidean}
\alias{euclidean,DepMapGeneExpression-method}
\title{Calculate Euclidean distance}
\usage{
euclidean(x, ...)

\S4method{euclidean}{DepMapGeneExpression}(x, upGenes, downGenes)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{up}{list of genes to be up-regulated}

\item{down}{list of genes to be down-regulated}
}
\value{
Numeric vector.
}
\description{
Calculate Euclidean distance
}
\note{
Updated 2023-09-11.
}
\examples{
showMethods("euclidean")
}
