% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/excludeContaminatedCells-methods.R
\name{excludeContaminatedCells}
\alias{excludeContaminatedCells}
\alias{excludeContaminatedCells,DepMapExperiment-method}
\title{Exclude contaminated cells}
\usage{
excludeContaminatedCells(object, ...)

\S4method{excludeContaminatedCells}{DepMapExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object, with contaminated cells excluded (removed).
}
\description{
Exclude contaminated cells
}
\note{
Updated 2023-09-12.
}
\examples{
data(crispr)

## DepMapExperiment ====
object <- crispr
object <- excludeContaminatedCells(object)
print(object)
}
\seealso{
\itemize{
\item Cellosaurus package.
}
}
