% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/filterProblematic-methods.R
\name{filterProblematic}
\alias{filterProblematic}
\alias{filterProblematic,DepMapGeneEffect-method}
\title{Filter problematic cell lines}
\usage{
filterProblematic(object, ...)

\S4method{filterProblematic}{DepMapGeneEffect}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with problematic cell lines removed.
}
\description{
Filter problematic cell lines
}
\note{
Updated 2023-08-22.
}
\examples{
data(crispr)

## DepMapExperiment ====
object <- crispr
object <- filterProblematic(object)
print(object)
}
