% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneEffect.R
\name{GeneEffect}
\alias{GeneEffect}
\title{Gene effect in cancer cell lines}
\usage{
GeneEffect(dataset = c("depmap_public_22q2", "demeter2_data_v6"))
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{GeneEffect}.
}
\description{
Gene effect in cancer cell lines
}
\note{
Updated 2022-08-05.
}
\section{Assays}{

\itemize{
\item \code{effect}: \strong{Chronos or CERES data} with principle components strongly
related to known batch effects removed, then shifted and scaled per cell
line so the median nonessential KO effect is 0 and the median essential KO
effect is -1.
\item \code{probability}: \strong{Probability} that knocking out the gene has a real
depletion effect using \code{gene_effect}.
}
}

\examples{
## CRISPR (default).
object <- GeneEffect()
print(object)

## RNAi.
## > object <- GeneEffect(dataset = "demeter2_data_v6")
}
