% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/humanize-methods.R
\name{humanize}
\alias{humanize}
\alias{humanize,DepMapGeneEffect-method}
\title{Humanize an R object}
\usage{
humanize(object, ...)

\S4method{humanize}{DepMapGeneEffect}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains human-friendly rownames (e.g. gene symbols instead of stable gene
IDs) and colnames (e.g. sample names instead of sample IDs).
}
\description{
Humanize an R object
}
\details{
This can make dimnames invalid (see \code{\link[base:make.names]{make.names()}}) and
should only be called prior to writing files to disk.
}
\note{
Updated 2023-03-20.
}
\examples{
data(crispr)

## DepMapGeneEffect ====
object <- crispr
print(head(rownames(object)))
print(head(colnames(object)))
object <- humanize(object)
print(head(rownames(object)))
print(head(colnames(object)))
}
