% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importAchillesGeneDependencyData}
\alias{importAchillesGeneDependencyData}
\title{Import Achilles gene dependency data}
\usage{
importAchillesGeneDependencyData(release = NULL)
}
\value{
\code{matrix}.
}
\description{
Import Achilles gene dependency data
}
\note{
Updated 2020-09-30.
}
\examples{
mat <- importAchillesGeneDependencyData()
}
