% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importAchillesGeneEffectData}
\alias{importAchillesGeneEffectData}
\title{Import Achilles gene effect data}
\usage{
importAchillesGeneEffectData(release = NULL)
}
\value{
\code{matrix}.
}
\description{
Import Achilles gene effect data
}
\note{
Updated 2020-09-30.
}
\examples{
mat <- importAchillesGeneEffectData()
}
