% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importCCLECopyNumberData}
\alias{importCCLECopyNumberData}
\title{Import CCLE copy number data}
\usage{
importCCLECopyNumberData(release = NULL)
}
\value{
\code{DataFrame}.
}
\description{
Import CCLE copy number data
}
\note{
Updated 2020-09-30.
}
\examples{
df <- importCCLECopyNumberData()
}
