% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCCLEExpressionData.R
\name{importCCLEExpressionData}
\alias{importCCLEExpressionData}
\title{Import CCLE expression data}
\usage{
importCCLEExpressionData(release = NULL)
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}
}
\value{
\code{DataFrame}.
}
\description{
Import CCLE expression data
}
\note{
Updated 2020-09-30.
}
\examples{
df <- importCCLEExpressionData()
dim(df)
}
