% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importCCLEMutationData}
\alias{importCCLEMutationData}
\title{Import CCLE mutation data}
\usage{
importCCLEMutationData(release = NULL)
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}
}
\value{
\code{DataFrame}.
}
\description{
Import CCLE mutation data
}
\note{
This file is large and takes a while to import.

Updated 2020-09-30.
}
\examples{
## This file is large and takes a while to import.
df <- importCCLEMutationData()
}
