% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCellLineSampleInfo.R
\name{importCellLineSampleInfo}
\alias{importCellLineSampleInfo}
\title{Import cell line sample info}
\usage{
importCellLineSampleInfo(release = NULL)
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}
}
\value{
\code{DataFrame}.
}
\description{
Import cell line sample info
}
\note{
Updated 2020-09-30.
}
\examples{
df <- importCellLineSampleInfo()
dim(df)
}
