% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDEMETER2GeneEffectData.R
\name{importDEMETER2GeneEffectData}
\alias{importDEMETER2GeneEffectData}
\title{Import DEMETER2 RNAi screen gene effect data}
\usage{
importDEMETER2GeneEffectData()
}
\value{
\code{matrix}.
}
\description{
Import DEMETER2 RNAi screen gene effect data
}
\note{
Updated 2020-09-30.
}
\examples{
mat <- importDEMETER2GeneEffectData()
dim(mat)
}
