% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCells-methods.R
\name{mapCells}
\alias{mapCells}
\alias{mapCells,Achilles-method}
\title{Map input to cell identifiers}
\usage{
mapCells(object, ...)

\S4method{mapCells}{Achilles}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character}.
}
\description{
Map input to cell identifiers
}
\note{
Updated 2020-10-01.
}
\examples{
class(mapCells)
}
