% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/matchGenes-methods.R
\name{matchGenes}
\alias{matchGenes}
\alias{matchGenes,AchillesGeneDependencyData-method}
\title{Match genes}
\usage{
matchGenes(object, ...)

\S4method{matchGenes}{AchillesGeneDependencyData}(object, GRanges, ids)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{GRanges}{\code{GRanges}.
Must contain Entrez identifiers defined in \code{entrezID} column of \code{mcols}.}

\item{ids}{\code{character}.
Ensembl gene identifiers, corresponding to the \code{names} of the \code{GRanges}
that contains the Entrez identifiers.}
}
\description{
Match genes
}
\note{
Updated 2020-09-30.
}
\examples{
object <- AchillesGeneDependencyData()
GRanges <- makeGRangesFromEnsembl("Homo sapiens")
ids <- head(names(GRanges))
matchGenes(
    object = object,
    GRanges = GRanges,
    ids = ids
)
}
\seealso{
\itemize{
\item \code{basejump::makeGRangesFromEnsembl}, for generating \code{GenomicRanges}.
}
}
