% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGeneEffect-methods.R
\name{plotGeneEffect}
\alias{plotGeneEffect}
\alias{plotGeneEffect,DepMapGeneEffect-method}
\title{Plot gene effect}
\usage{
plotGeneEffect(object, ...)

\S4method{plotGeneEffect}{DepMapGeneEffect}(object, genes, geom = c("boxplot", "violin", "density"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}
}
\value{
\code{ggplot}.
}
\description{
Plot gene effect
}
\note{
Updated 2023-01-27.
}
\examples{
data(crispr)

## DepMapGeneEffect ====
object <- crispr
genes <- head(rownames(object), n = 5L)
plotGeneEffect(object, genes = genes)
}
