% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/predictSensitivity-methods.R
\name{predictSensitivity}
\alias{predictSensitivity}
\alias{predictSensitivity,DepMapGeneExpression-method}
\title{Predict sensitivity}
\usage{
predictSensitivity(object, ...)

\S4method{predictSensitivity}{DepMapGeneExpression}(object, upregulated, downregulated)
}
\arguments{
\item{object}{Object.}

\item{upregulated}{\code{character}.
Genes observed to be upregulated by treatment.}

\item{downregulated}{\code{character}.
Genes observed to be downregulated by treatment.}
}
\value{
\code{DFrame}.
}
\description{
Predict sensitivity
}
\details{
Predict cell line sensitivity by calculation of Euclidean distance between
known biomarker genes up- or down-regulated by a treatment of interest.
}
\note{
Updated 2023-09-12.
}
\examples{
data(rnaseq)

## DepMapGeneExpression ====
object <- rnaseq
genes <- rownames(object)
upregulated <- genes[1L:5L]
downregulated <- genes[6L:10L]
df <- predictSensitivity(
    object = object,
    upregulated = upregulated,
    downregulated = downregulated
)
print(df)
}
\seealso{
\itemize{
\item DepMapAnalysis package.
}
}
\author{
Nitesh Turaga, Teresa Rice, Michael Steinbaugh
}
