% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnai}
\alias{rnai}
\title{RNAi example data set}
\format{
An object of class \code{GeneEffect} with 100 rows and 50 columns.
}
\usage{
data(rnai)
}
\value{
\code{DESeqAnalysis}.
}
\description{
RNAi example data set
}
\note{
Updated 2021-06-09.
}
\examples{
data(rnai)
print(rnai)
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
