% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaseq}
\alias{rnaseq}
\title{RNA-seq example data set}
\format{
An object of class \code{CCLEExpressionData} with 100 rows and 50 columns.
}
\usage{
data(rnaseq)
}
\value{
\code{CCLEExpressionData}.
}
\description{
RNA-seq example data set
}
\note{
Updated 2022-08-05.
}
\examples{
data(rnaseq)
print(rnaseq)
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
