% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/selectCells-methods.R
\name{selectCells}
\alias{selectCells}
\alias{selectCells,DepMapGeneEffect-method}
\title{Select cells}
\usage{
selectCells(object, ...)

\S4method{selectCells}{DepMapGeneEffect}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Key value pairs that map to cell line metadata.}
}
\value{
Modified object.
}
\description{
Select cells
}
\details{
Only exact matching is currently supported.
}
\note{
Updated 2023-09-14.
}
\examples{
data(crispr)

## DepMapGeneEffect ====
object <- crispr
print(dim(object))
oncotreeCode <-
    colData(object) |>
    _[["cellosaurus"]] |>
    _[["oncotreeCode"]] |>
    decode() |>
    _[[1L]]
print(oncotreeCode)
subset <- selectCells(object, oncotreeCode = oncotreeCode)
print(dim(object))
}
