% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show}
\alias{show}
\alias{show,DepMapAnalysis-method}
\title{Show an object}
\usage{
\S4method{show}{DepMapAnalysis}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Console output.
}
\description{
Show an object
}
\note{
Updated 2021-06-09.
}
\examples{
data(crispr)

## DepMapAnalysis ====
show(crispr)
}
