% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tpm-methods.R
\name{tpm}
\alias{tpm}
\alias{tpm,DepMapGeneExpression-method}
\title{Transcripts per million (log2)}
\usage{
tpm(object, ...)

\S4method{tpm}{DepMapGeneExpression}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}
Log2 transcripts per million.
Cells in columns, genes in rows.
}
\description{
Transcripts per million
}
\note{
Updated 2023-09-11.
}
\examples{
data(rnaseq)
tpm <- tpm(rnaseq)
print(tpm[1L:5L, 1L:5L])
}
