% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zscore-methods.R
\name{zscore}
\alias{zscore}
\alias{zscore,DepMapGeneExpression-method}
\title{Calculate Z-score}
\usage{
zscore(object, ...)

\S4method{zscore}{DepMapGeneExpression}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}
Z-score normalized log2 transcripts per million.
Cells in columns, genes in rows.
}
\description{
Calculate Z-score
}
\note{
Updated 2023-09-11.
}
\examples{
data(rnaseq)
zscore <- zscore(rnaseq)
print(zscore[1L:5L, 1L:5L])
}
\seealso{
\itemize{
\item \code{scale()}.
\item \code{limma::zscore()}.
}
}
