% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ach}
\alias{ach}
\title{Achilles example data set}
\format{
An object of class \code{Achilles} of length 1.
}
\usage{
ach
}
\description{
Achilles example data set
}
\note{
Updated 2021-02-24.
}
\examples{
data(ach)
ach
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
