% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Achilles-class}
\alias{Achilles-class}
\title{Achilles CRISPR screening data}
\value{
\code{Achilles}.
}
\description{
Achilles CRISPR screening data
}
\details{
Scored using CERES algorithm.

Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2021-02-25.
}
\seealso{
\itemize{
\item https://depmap.org/portal/achilles/
\item https://depmap.org/ceres/
}
}
