% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{AchillesGeneDependencyData}
\alias{AchillesGeneDependencyData}
\title{Import Achilles gene dependency data}
\usage{
AchillesGeneDependencyData(release = NULL)
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}
}
\value{
\code{AchillesGeneDependencyData}.
}
\description{
Import Achilles gene dependency data
}
\note{
Updated 2020-09-30.
}
\examples{
object <- AchillesGeneDependencyData()
dim(object)
}
