% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{AchillesGeneEffectData}
\alias{AchillesGeneEffectData}
\title{Import Achilles gene effect data}
\usage{
AchillesGeneEffectData(release = NULL)
}
\value{
\code{AchillesGeneEffectData}.
}
\description{
Import Achilles gene effect data
}
\note{
Updated 2020-09-30.
}
\examples{
object <- AchillesGeneEffectData()
dim(object)
}
