% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{AchillesGeneEffectData}
\alias{AchillesGeneEffectData}
\title{Import Achilles gene effect data}
\usage{
AchillesGeneEffectData(release = NULL)
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}
}
\value{
\code{AchillesGeneEffectData}.
}
\description{
\strong{CERES data} with principle components strongly related to known batch
effects removed, then shifted and scaled per cell line so the median
nonessential KO effect is 0 and the median essential KO effect is -1.
}
\note{
Updated 2020-09-30.
}
\examples{
object <- AchillesGeneEffectData()
dim(object)
}
