% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CCLECopyNumberData-class}
\alias{CCLECopyNumberData-class}
\title{CCLE copy number data}
\value{
\code{CCLECopyNumberData}.
}
\description{
CCLE copy number data
}
\details{
Gene level copy number data, log2 transformed with a pseudo count of 1.
This is generated by mapping genes onto the segment level calls.

Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2021-06-09.
}
