% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE.R
\name{CCLECopyNumberData}
\alias{CCLECopyNumberData}
\title{Import CCLE copy number data}
\usage{
CCLECopyNumberData(
  dataset = "depmap_public_21q2",
  rowData = TRUE,
  colData = TRUE
)
}
\arguments{
\item{dataset}{\code{character(1)} or \code{NULL}.
DepMap dataset release name.
If left \code{NULL}, defaults to latest supported dataset.}

\item{rowData}{\code{logical(1)}.
Include gene annotations as genomic ranges.}

\item{colData}{\code{logical(1)}.
Include cell line metadata.}
}
\value{
\code{CCLECopyNumberData}.
}
\description{
Import CCLE copy number data
}
\note{
Updated 2021-02-25.
}
\examples{
object <- CCLECopyNumberData()
print(object)
}
