% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{CCLECopyNumberData}
\alias{CCLECopyNumberData}
\title{Import CCLE copy number data}
\usage{
CCLECopyNumberData(release = NULL)
}
\value{
\code{CCLECopyNumberData}.
}
\description{
Import CCLE copy number data
}
\note{
Updated 2020-09-30.
}
\examples{
object <- CCLECopyNumberData()
dim(object)
}
