% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLECopyNumberData.R
\name{CCLECopyNumberData}
\alias{CCLECopyNumberData}
\title{Import CCLE copy number data}
\usage{
CCLECopyNumberData(dataset = "depmap_public_22q2")
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{CCLECopyNumberData}.
}
\description{
Import CCLE copy number data
}
\note{
Updated 2021-07-07.
}
\examples{
object <- CCLECopyNumberData()
print(object)
}
