% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLECopyNumberData.R
\name{CCLECopyNumberData}
\alias{CCLECopyNumberData}
\title{Import CCLE copy number data}
\usage{
CCLECopyNumberData(
  dataset = c("depmap_public_22q1", "depmap_public_21q4", "depmap_public_21q3",
    "depmap_public_21q2", "depmap_public_21q1")
)
}
\arguments{
\item{dataset}{\code{character(1)}.
DepMap dataset release name.}
}
\value{
\code{CCLECopyNumberData}.
}
\description{
Import CCLE copy number data
}
\note{
Updated 2021-07-07.
}
\examples{
object <- CCLECopyNumberData()
print(object)
}
