% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CCLEExpressionData-class}
\alias{CCLEExpressionData-class}
\title{CCLE expression data}
\value{
\code{CCLEExpressionData}.
}
\description{
CCLE expression data
}
\details{
RNA-seq TPM gene expression data for just protein coding genes using RSEM.
Log2 transformed, using a pseudo-count of 1.

Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2022-08-05.
}
