% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CCLEExpressionData-class}
\alias{CCLEExpressionData-class}
\title{CCLE expression data}
\value{
\code{CCLEExpressionData}.
}
\description{
CCLE expression data
}
\details{
Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2021-02-25.
}
