% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE.R
\name{CCLEExpressionData}
\alias{CCLEExpressionData}
\title{Import CCLE expression data}
\usage{
CCLEExpressionData(release = "21Q1")
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}
}
\value{
\code{CCLEExpressionData}.
}
\description{
Import CCLE expression data
}
\note{
Updated 2021-02-24.
}
\examples{
object <- CCLEExpressionData()
dim(object)
}
