% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEExpressionData.R
\name{CCLEExpressionData}
\alias{CCLEExpressionData}
\title{Import CCLE expression data}
\usage{
CCLEExpressionData(
  dataset = c("depmap_public_22q2", "depmap_public_22q1", "depmap_public_21q4",
    "depmap_public_21q3", "depmap_public_21q2", "depmap_public_21q1")
)
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{CCLEExpressionData}.
}
\description{
Import CCLE expression data
}
\note{
Updated 2021-07-07.
}
\examples{
object <- CCLEExpressionData()
dim(object)
}
