% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEExpressionData.R
\name{CCLEExpressionData}
\alias{CCLEExpressionData}
\title{Import CCLE expression data}
\usage{
CCLEExpressionData(
  dataset = c("depmap_public_21q2", "depmap_public_21q1", "depmap_public_20q4v2",
    "depmap_public_20q3", "depmap_public_20q2", "depmap_public_20q1")
)
}
\arguments{
\item{dataset}{\code{character(1)} or \code{NULL}.
DepMap dataset release name.
If left \code{NULL}, defaults to latest supported dataset.}
}
\value{
\code{CCLEExpressionData}.
}
\description{
Import CCLE expression data
}
\note{
Updated 2021-07-07.
}
\examples{
object <- CCLEExpressionData()
dim(object)
}
