% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CCLEFusionData-class}
\alias{CCLEFusionData-class}
\title{CCLE fusion data}
\value{
\code{CCLEFusionData}.
}
\description{
CCLE fusion data
}
\details{
Inherits from \code{DataFrame}.
}
\note{
Updated 2021-02-25.
}
