% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEMutationData.R
\name{CCLEMutationData}
\alias{CCLEMutationData}
\title{Import CCLE mutation data}
\usage{
CCLEMutationData(dataset = "depmap_public_22q2")
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{CCLEMutationData}.
}
\description{
Import CCLE mutation data
}
\note{
Updated 2022-09-21.
}
\examples{
object <- CCLEMutationData()
dim(object)
}
