% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEMutationData.R
\name{CCLEMutationData}
\alias{CCLEMutationData}
\title{Import CCLE mutation data}
\usage{
CCLEMutationData(
  dataset = c("depmap_public_21q2", "depmap_public_21q1", "depmap_public_20q4v2",
    "depmap_public_20q3", "depmap_public_20q2", "depmap_public_20q1")
)
}
\arguments{
\item{dataset}{\code{character(1)} or \code{NULL}.
DepMap dataset release name.
If left \code{NULL}, defaults to latest supported dataset.}
}
\value{
\code{CCLEMutationData}.
}
\description{
Import CCLE mutation data
}
\note{
Updated 2021-07-07.
}
\examples{
object <- CCLEMutationData()
dim(object)
}
