% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE.R
\name{CCLEMutationData}
\alias{CCLEMutationData}
\title{Import CCLE mutation data}
\usage{
CCLEMutationData(dataset = "depmap_public_21q2")
}
\arguments{
\item{dataset}{\code{character(1)} or \code{NULL}.
DepMap dataset release name.
If left \code{NULL}, defaults to latest supported dataset.}
}
\value{
\code{CCLEMutationData}.
}
\description{
Import CCLE mutation data
}
\note{
Updated 2021-02-25.
}
\examples{
object <- CCLEMutationData()
dim(object)
}
