% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEMutationData.R
\name{CCLEMutationData}
\alias{CCLEMutationData}
\title{Import CCLE mutation data}
\usage{
CCLEMutationData(
  dataset = c("depmap_public_22q2", "depmap_public_22q1", "depmap_public_21q4",
    "depmap_public_21q3", "depmap_public_21q2", "depmap_public_21q1")
)
}
\arguments{
\item{dataset}{\code{character(1)}.
DepMap dataset release name.}
}
\value{
\code{CCLEMutationData}.
}
\description{
Import CCLE mutation data
}
\note{
Updated 2022-03-09.
}
\examples{
object <- CCLEMutationData()
dim(object)
}
