% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codependencies.R
\name{Codependencies}
\alias{Codependencies}
\title{Import Co-dependency scores}
\usage{
Codependencies(
  gene,
  dataset = c("Chronos_Combined", "CERES_Combined", "RNAi_merged")
)
}
\arguments{
\item{gene}{\code{character(1)}.
Gene identifier.}

\item{dataset}{\code{character(1)}.
DepMap dataset release name.}
}
\value{
\code{Codependencies}.
}
\description{
Import Co-dependency scores
}
\note{
Updated 2021-07-19.
}
\examples{
x <- Codependencies(gene = "SOX10")
}
