% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codependencies.R
\name{Codependencies}
\alias{Codependencies}
\title{Import Co-dependency scores}
\usage{
Codependencies(
  geneName,
  dataset = c("Chronos_Combined", "CERES_Combined", "RNAi_merged")
)
}
\arguments{
\item{geneName}{\code{character(1)}.
Gene name (a.k.a. symbol).}

\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{Codependencies}.
}
\description{
Import Co-dependency scores
}
\note{
Updated 2022-08-05.
}
\examples{
x <- Codependencies(geneName = "SOX10")
}
\seealso{
\itemize{
\item https://forum.depmap.org/t/down-load-of-co-dependencies/175/
\item Python script for processing of all co-dependency scores:
https://gist.github.com/pgm/ac2ac4c664ef81200ce49133cc4cee02
}
}
