% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crispr}
\alias{crispr}
\title{CRISPR example data set}
\format{
An object of class \code{DepMapAnalysis} with 100 rows and 50 columns.
}
\usage{
data(crispr)
}
\description{
CRISPR example data set
}
\note{
Updated 2021-06-09.
}
\examples{
data(crispr)
crispr
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
