% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapAnalysis.R
\name{DepMapAnalysis}
\alias{DepMapAnalysis}
\title{Import cancer cell line dependency map data}
\usage{
DepMapAnalysis(
  release = "21Q2",
  scoringMethod = c("chronos", "ceres"),
  rowData = TRUE,
  colData = TRUE
)
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
DepMap quarterly release ID (e.g. \code{"20Q3"}).
If left \code{NULL}, defaults to latest supported release.}

\item{rowData}{\code{logical(1)}.
Include gene annotations as genomic ranges.}

\item{colData}{\code{logical(1)}.
Include cell line metadata.}
}
\value{
\code{DepMapAnalysis}.
}
\description{
Import cancer cell line dependency map data
}
\note{
Updated 2021-06-09.
}
\section{Assays}{

\itemize{
\item \code{effect}: \strong{CERES data} with principle components strongly related to
known batch effects removed, then shifted and scaled per cell line so the
median nonessential KO effect is 0 and the median essential KO effect is
-1.
\item \code{probability}: \strong{Probability} that knocking out the gene has a real
depletion effect using \code{gene_effect}.
}
}

\examples{
object <- Achilles()
print(object)
}
