% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/DepMapCodependencies-methods.R
\name{DepMapCodependencies}
\alias{DepMapCodependencies}
\alias{DepMapCodependencies,DepMapGeneEffect}
\alias{DepMapCodependencies,DepMapGeneEffect-method}
\title{Calculate co-dependency scores}
\usage{
DepMapCodependencies(object, ...)

`DepMapCodependencies,DepMapGeneEffect`(object, gene1, gene2 = NULL, ...)

\S4method{DepMapCodependencies}{DepMapGeneEffect}(object, gene1, gene2 = NULL, ...)
}
\arguments{
\item{geneName1}{\code{character(1)}.
Gene name.}

\item{geneName2}{\code{character(1)} or \code{NULL}.
Gene name or \code{NULL}.
If \code{NULL}, calculate correlations against all genes.}
}
\value{
\code{DepMapCodependencies}.
}
\description{
Calculate co-dependency scores
}
\note{
Updated 2023-01-27.
}
\examples{
data(crispr)

## DepMapGeneEffect ====
object <- crispr

geneNames <- as.character(rowData(object)[["geneName"]])
geneName1 <- geneNames[[1L]]
geneName2 <- geneNames[[2L]]

## Calculate all co-dependencies for a gene of interest.
x <- DepMapCodependencies(
    object = object,
    geneName1 = geneName1,
    geneName2 = NULL
)
print(head(x))

## Lineage restrict and compare 2 genes.
diseases <- as.character(colData(object)[["cellosaurusNcItDisease"]])
disease <- diseases[[1L]]
keep <- colData(object)[["cellosaurusNcItDisease"]] \%in\% disease
object <- object[, keep]
x <- DepMapCodependencies(
    object = object,
    geneName1 = geneName1,
    geneName2 = geneName2,
    lineage = "pancreas"
)
print(x)
}
\seealso{
\itemize{
\item \href{https://forum.depmap.org/t/down-load-of-co-dependencies/175/}{How to download codependency scores}
\item \href{https://gist.github.com/pgm/ac2ac4c664ef81200ce49133cc4cee02}{Python script for processing of all codependency scores}
}
}
