% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapCopyNumber.R
\name{DepMapCopyNumber}
\alias{DepMapCopyNumber}
\title{Import DepMap (CCLE) copy number data}
\usage{
DepMapCopyNumber(dataset = "depmap_public_22q2")
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{DepMapCopyNumber}.
}
\description{
Import DepMap (CCLE) copy number data
}
\note{
Updated 2023-01-26.
}
\examples{
object <- DepMapCopyNumber()
print(object)
}
