% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapCopyNumber.R
\name{DepMapCopyNumber}
\alias{DepMapCopyNumber}
\title{Import DepMap copy number data}
\usage{
DepMapCopyNumber()
}
\value{
\code{DepMapCopyNumber}.
}
\description{
Import DepMap copy number data
}
\note{
Updated 2023-08-08.
}
\examples{
object <- DepMapCopyNumber()
print(object)
}
