% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapExperiment-class}
\alias{DepMapExperiment-class}
\title{DepMap experiment}
\value{
\code{DepMapExperiment}.
}
\description{
DepMap experiment
}
\details{
Virtual class, not intended to be worked with directly.

Extends \code{SummarizedExperiment} class, with additional DepMap-specific
validity checks.
}
\note{
Updated 2023-08-22.
}
