% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapExpression-class}
\alias{DepMapExpression-class}
\title{DepMap RNA-seq expression data}
\value{
\code{DepMapExpression}.
}
\description{
DepMap RNA-seq expression data
}
\details{
RNA-seq TPM gene expression data for just protein coding genes using RSEM.
Log2 transformed, using a pseudo-count of 1.

Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.
}
\note{
Updated 2023-01-27.
}
