% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapExpression.R
\name{DepMapExpression}
\alias{DepMapExpression}
\title{Import DepMap expression data}
\usage{
DepMapExpression()
}
\value{
\code{DepMapExpression}.
}
\description{
Import DepMap expression data
}
\note{
Updated 2023-08-08.
}
\examples{
object <- DepMapExpression()
dim(object)
}
