% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapExpression.R
\name{DepMapExpression}
\alias{DepMapExpression}
\title{Import DepMap expression data}
\usage{
DepMapExpression(dataset = c("depmap_public_22q4", "depmap_public_22q2"))
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{DepMapExpression}.
}
\description{
Import DepMap expression data
}
\note{
Updated 2022-01-26.
}
\examples{
object <- DepMapExpression()
dim(object)
}
