% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapGeneEffect-class}
\alias{DepMapGeneEffect-class}
\title{DepMap \code{GeneEffect} class}
\value{
\code{DepMapGeneEffect}.
}
\description{
DepMap \code{GeneEffect} class
}
\details{
Inherits from \code{SummarizedExperiment}.
Cells in columns, genes in rows.

Not intended to be used directly.

\code{DepMapCrisprGeneEffect} and \code{DepMapRnaiGeneEffect} extend this class.
}
\note{
Updated 2023-08-09.
}
