% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapGeneEffect-class}
\alias{DepMapGeneEffect-class}
\title{\code{DepMapGeneEffect} virtual class}
\value{
\code{DepMapGeneEffect}.
}
\description{
\code{DepMapGeneEffect} virtual class
}
\details{
Not intended to be used directly.

\code{DepMapCrisprGeneEffect} and \code{DepMapRnaiGeneEffect} extend this class.
}
\note{
Updated 2023-09-14.
}
