% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapGeneEffect.R
\name{DepMapGeneEffect}
\alias{DepMapGeneEffect}
\title{DepMap gene effect in cancer cell lines}
\usage{
DepMapGeneEffect(dataset = c("depmap_public_23q2", "demeter2_data_v6"))
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{DepMapGeneEffect}.
}
\description{
DepMap gene effect in cancer cell lines
}
\note{
Updated 2023-03-08.
}
\section{Assays}{

\itemize{
\item \code{effect}: \strong{Chronos or CERES data} with principle components strongly
related to known batch effects removed, then shifted and scaled per cell
line so the median nonessential KO effect is 0 and the median essential KO
effect is -1.
\item \code{probability}: \strong{Probability} that knocking out the gene has a real
depletion effect using \code{gene_effect}.
}
}

\examples{
## CRISPR (default).
object <- DepMapGeneEffect()
print(object)

## RNAi (DEMETER2).
## > object <- DepMapGeneEffect(dataset = "demeter2_data_v6")
}
