% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapMicroRNA-XXX.R
\name{DepMapMicroRna}
\alias{DepMapMicroRna}
\title{Import DepMap microRNA expression data}
\usage{
DepMapMicroRna()
}
\value{
\code{DepMapMicroRna}.
}
\description{
Import DepMap microRNA expression data
}
\details{
NanoString microRNA panel data.
}
\note{
Updated 2023-08-09.
}
\examples{
object <- DepMapMicroRna()
dim(object)
}
