% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepMapMutations.R
\name{DepMapMutations}
\alias{DepMapMutations}
\title{Import DepMap somatic mutation data}
\usage{
DepMapMutations(dataset = "depmap_public_23q2")
}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}
}
\value{
\code{DepMapMutations}.
}
\description{
Import DepMap somatic mutation data
}
\note{
Updated 2023-08-02.
}
\examples{
object <- DepMapMutations()
dim(object)
}
