% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLEProteomicsData.R
\name{DepMapProteomics}
\alias{DepMapProteomics}
\title{Import DepMap proteomics data}
\usage{
DepMapProteomics()
}
\value{
\code{DepMapProteomics}.
}
\description{
Import DepMap proteomics data
}
\note{
Updated 2023-01-27.
}
\seealso{
\itemize{
\item https://depmap.org/portal/download/all/?releasename=Proteomics
\item https://gygi.hms.harvard.edu/publications/ccle.html
}
}
