% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapRnaiGeneEffect-class}
\alias{DepMapRnaiGeneEffect-class}
\title{DepMap RNAi gene effect (DEMETER2)}
\value{
\code{DepMapRnaiGeneEffect}.
}
\description{
DepMap RNAi gene effect (DEMETER2)
}
\details{
Inherits from \code{DepMapGeneEffect}, which extends \code{SummarizedExperiment}.
Cells in columns, NCBI genes in rows.
}
\note{
Updated 2023-08-09.
}
