% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DepMapTxExpression-class}
\alias{DepMapTxExpression-class}
\title{DepMap RNA-seq transcript expression}
\value{
\code{DepMapTxExpression}.
}
\description{
DepMap RNA-seq transcript expression
}
\details{
Inherits from \code{DepMapExpression}, which extends \code{SummarizedExperiment}.
Cells in columns, transcripts in rows.
}
\note{
Updated 2023-08-09.
}
