% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/euclidean-methods.R
\name{euclidean}
\alias{euclidean}
\alias{euclidean,DepMapGeneExpression-method}
\title{Calculate Euclidean distance}
\usage{
euclidean(x, ...)

\S4method{euclidean}{DepMapGeneExpression}(x, upregulated, downregulated)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{upregulated}{\code{character}.
Genes observed to be upregulated by treatment.}

\item{downregulated}{\code{character}.
Genes observed to be downregulated by treatment.}
}
\value{
Numeric vector.
}
\description{
Calculate Euclidean distance
}
\note{
Updated 2023-09-11.
}
\examples{
rnaseq <- DepMapGeneExpression()
}
