% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/findBiomarkers-methods.R
\name{findBiomarkers}
\alias{findBiomarkers}
\alias{findBiomarkers,DepMapCopyNumber-method}
\alias{findBiomarkers,DepMapExpression-method}
\alias{findBiomarkers,DepMapMutation-method}
\title{Find biomarkers}
\usage{
findBiomarkers(object, ...)

\S4method{findBiomarkers}{DepMapCopyNumber}(object, sensitiveCells, insensitiveCells)

\S4method{findBiomarkers}{DepMapExpression}(object, sensitiveCells, insensitiveCells)

\S4method{findBiomarkers}{DepMapMutation}(object, sensitiveCells, insensitiveCells)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{insensitiveCells, sensitiveCells}{\code{character}.
Cell line names.}
}
\value{
FIXME Varies, depending on method.
}
\description{
Find biomarkers
}
\note{
Updated 2023-01-27.
}
\examples{
data(rnaseq)

## DepMapExpression ====
object <- rnaseq
sensitiveCells <- colnames(object)[c(1L:5L)]
insensitiveCells <- colnames(object)[c(6L:10L)]
x <- findBiomarkers(
    object = object,
    sensitiveCells = sensitiveCells,
    insensitiveCells = insensitiveCells
)
print(x)
}
\seealso{
\itemize{
\item limma User's Guide, Section 15.4.
}
}
