% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneEffect.R
\name{GeneEffect}
\alias{GeneEffect}
\title{Gene effect in cancer cell lines}
\usage{
GeneEffect(
  dataset = c("depmap_public_21q2", "depmap_public_21q1", "depmap_public_20q4v2",
    "depmap_public_20q3", "depmap_public_20q2", "depmap_public_20q1",
    "sanger_project_score_2021_05", "sanger_project_score_2019_08", "demeter2_data_v6"),
  project = "default",
  scoringMethod = "default"
)
}
\arguments{
\item{dataset}{\code{character(1)}.
DepMap dataset release name.}

\item{project}{\code{character(1)}.
Project name.
Defaults to a combination of multiple projects (\code{"combined"}):
\itemize{
\item CRISPR: Broad DepMap Public, Sanger Project Score
(e.g. for \code{depmap_public_21q2}).
\item RNAi: Achilles, DRIVE, Marcotte
(e.g. for \code{demeter2_data_v6}).
}}

\item{scoringMethod}{\code{character(1)}.
Scoring method name to use.
\itemize{
\item CRISPR: \code{"chronos"} (as of 2021 Q1) or \code{"ceres"}
(e.g. \code{depmap_public_21q2} dataset).
\item RNAi: \code{"demeter2"}
(e.g. \code{demeter2_data_v6} dataset).
}}
}
\value{
\code{GeneEffect}.
}
\description{
Gene effect in cancer cell lines
}
\note{
Updated 2021-07-08.
}
\section{Assays}{

\itemize{
\item \code{effect}: \strong{Chronos or CERES data} with principle components strongly
related to known batch effects removed, then shifted and scaled per cell
line so the median nonessential KO effect is 0 and the median essential KO
effect is -1.
\item \code{probability}: \strong{Probability} that knocking out the gene has a real
depletion effect using \code{gene_effect}.
}
}

\examples{
object <- GeneEffect()
print(object)
}
