% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenes-methods.R
\name{mapGenes}
\alias{mapGenes}
\alias{mapGenes,Achilles-method}
\title{Map input to gene identifiers}
\usage{
mapGenes(object, ...)

\S4method{mapGenes}{Achilles}(object, GRanges, ids)
}
\arguments{
\item{object}{Object.}

\item{GRanges}{\code{GRanges}.
Must contain Entrez identifiers defined in \code{entrezID} column of \code{mcols}.}

\item{ids}{\code{character}.
Ensembl gene identifiers, corresponding to the \code{names} of the \code{GRanges}
that contains the Entrez identifiers.}
}
\value{
\code{character}.
}
\description{
Map input to gene identifiers
}
\note{
Updated 2020-10-01.
}
\examples{
object <- AchillesGeneDependencyData()
GRanges <- makeGRangesFromEnsembl("Homo sapiens")
ids <- head(names(GRanges))
mapGenes(
    object = object,
    GRanges = GRanges,
    ids = ids
)
}
\seealso{
\itemize{
\item \code{basejump::makeGRangesFromEnsembl}, for generating \code{GenomicRanges}.
}
}
