% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{colData}{\code{logical(1)}.
Include cell line metadata.}

\item{dataset}{\code{character(1)} or \code{NULL}.
DepMap dataset release name.
If left \code{NULL}, defaults to latest supported dataset.}

\item{rowData}{\code{logical(1)}.
Include gene annotations as genomic ranges.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\keyword{internal}
