% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{dataset}{\code{character(1)} or \code{NULL}.
DepMap dataset release name.
If left \code{NULL}, defaults to latest supported dataset.}

\item{gene}{\code{character(1)}.
Gene identifier.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\keyword{internal}
