% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{dataset}{\code{character(1)}.
Dataset name.}

\item{gene}{\code{character(1)}.
Gene identifier.}

\item{geneName}{\code{character(1)}.
Gene name (a.k.a. symbol).}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\keyword{internal}
