% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotGeneEffectVsExpression-methods.R
\name{plotGeneEffectVsExpression}
\alias{plotGeneEffectVsExpression}
\alias{plotGeneEffectVsExpression,GeneEffect,CCLEExpressionData-method}
\title{Plot gene effect vs. expression}
\usage{
plotGeneEffectVsExpression(effect, expression, ...)

\S4method{plotGeneEffectVsExpression}{GeneEffect,CCLEExpressionData}(
  effect,
  expression,
  gene,
  subtype = NULL,
  subtypeCol = "subtype",
  label = FALSE
)
}
\arguments{
\item{effect}{\code{GeneEffect}.}

\item{expression}{\code{CCLEExpressionData}.}

\item{...}{Additional arguments.}

\item{gene}{\code{character(1)}.
Gene identifier.}

\item{subtype}{\code{character(1)} or \code{NULL}.
Cell subtype name to use for filtering.}

\item{subtypeCol}{\code{character(1)}.
Column name of metadata defined in \code{colData} of \code{effect}.}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}
}
\description{
Plot gene effect vs. expression
}
\note{
Updated 2021-07-15.
}
\examples{
data(crispr)

effect <- crispr
expression <- CCLEExpressionData()

## GeneEffect,CCLEExpressionData ====
gene <- rownames(effect)[[1L]]
plotGeneEffectVsExpression(
    effect = effect,
    expression = expression,
    gene = gene
)
}
