% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotTopGeneEffectPerCell-methods.R
\name{plotTopGeneEffectPerCell}
\alias{plotTopGeneEffectPerCell}
\alias{plotTopGeneEffectPerCell,DepMapAnalysis-method}
\title{Plot top gene effect (dependencies) per cell}
\usage{
plotTopGeneEffectPerCell(object, ...)

\S4method{plotTopGeneEffectPerCell}{DepMapAnalysis}(object, cells, n = 5L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{cells}{\code{character}.
Cell identifiers.}

\item{n}{\code{integer(1)}.
Number to include.}
}
\value{
\code{ggplot}.
}
\description{
Plot top gene effect (dependencies) per cell
}
\note{
Updated 2021-06-10.
}
\examples{
data(crispr)

## DepMapAnalysis ====
object <- crispr
cells <- head(colnames(object), n = 6L)
plotTopGeneEffectPerCell(object, cells = cells, n = 5L)
}
