% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotTopGeneEffectPerGroup-methods.R
\name{plotTopGeneEffectPerGroup}
\alias{plotTopGeneEffectPerGroup}
\alias{plotTopGeneEffectPerGroup,GeneEffect-method}
\title{Plot top gene effect per group of interest}
\usage{
plotTopGeneEffectPerGroup(object, ...)

\S4method{plotTopGeneEffectPerGroup}{GeneEffect}(
  object,
  gene,
  group = "subtype",
  n = 10L,
  minNPerGroup = 3L
)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{gene}{\code{character(1)}.
Gene identifier.}

\item{group}{\code{character(1)} or \code{NULL}.
Column name defined in \code{colData} to use for grouping.
(e.g. \code{"subtype"}).}

\item{n}{\code{integer(1)}.
Number to include.}

\item{minNPerGroup}{\code{integer(1)}.
Minimum number of samples per group to consider for inclusion in
plot.}
}
\value{
\code{ggplot}.
}
\description{
Plot top gene effect per group of interest
}
\note{
Updated 2021-07-07.
}
\examples{
data(crispr)

## GeneEffect ====
object <- crispr
gene <- rownames(object)[[1L]]
group <- "subtype"
plotTopGeneEffectPerGroup(object, gene = gene, group = group)
}
