% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/predictSensitivity-methods.R
\name{predictSensitivity}
\alias{predictSensitivity}
\title{Predict sensitivity}
\usage{
predictSensitivity(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{sensitive, insensitive}{\code{character}.
Cell line identifiers.}
}
\description{
Predict sensitivity
}
\note{
Updated 2023-07-03.
}
\examples{
data(rnaseq)

## DepMapExpression ====
object <- rnaseq
sensitive <- c("AAA", "BBB")
insensitive <- c("CCC", "DDD")
out <- predictSensitivity(
    object = object,
    sensitive = sensitive,
    insensitive = insensitive
)
print(out)
}
