% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnai}
\alias{rnai}
\title{RNAi example data set}
\format{
An object of class \code{GeneEffect} of length 1.
}
\usage{
data(rnai)
}
\value{
\code{DepMapGeneEffect}.
}
\description{
RNAi example data set
}
\note{
Updated 2023-01-26.
}
\examples{
data(rnai)
print(rnai)
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
