% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/selectCells-methods.R
\name{selectCells}
\alias{selectCells}
\alias{selectCells,CCLECopyNumberData-method}
\alias{selectCells,GeneEffect-method}
\title{Select cells}
\usage{
selectCells(object, ...)

\S4method{selectCells}{CCLECopyNumberData}(object, ...)

\S4method{selectCells}{GeneEffect}(object, ...)

\S4method{selectCells}{GeneEffect}(object, ...)
}
\description{
Select cells
}
\note{
Updated 2022-08-17.
}
\examples{
data(crispr)

## GeneEffect ====
object <- crispr
print(object)
lineage <- colData(object)[["lineage"]][[1L]]
x <- selectCells(object, lineage = lineage)
print(x)
}
\seealso{
\itemize{
\item \code{AcidExperiment::selectSamples}.
}
}
