% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/selectCells-methods.R
\name{selectCells}
\alias{selectCells}
\alias{selectCells,CCLECopyNumberData-method}
\alias{selectCells,CCLEExpressionData-method}
\alias{selectCells,GeneEffect-method}
\title{Select cells}
\usage{
selectCells(object, ...)

\S4method{selectCells}{CCLECopyNumberData}(object, ...)

\S4method{selectCells}{CCLEExpressionData}(object, ...)

\S4method{selectCells}{GeneEffect}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Key value pairs that map to cell line metadata defined in \code{colData}.}
}
\value{
Modified object.
}
\description{
Select cells
}
\note{
Updated 2022-08-17.
}
\examples{
data(crispr)

## GeneEffect ====
object <- crispr
dim(object)
lineage <- colData(object)[["lineage"]][[1L]]
print(lineage)
all(colData(object)[["lineage"]] == lineage)
object <- selectCells(object, lineage = lineage)
all(colData(object)[["lineage"]] == lineage)
dim(object)
}
\seealso{
\itemize{
\item \code{AcidExperiment::selectSamples}.
}
}
