% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/selectCells-methods.R
\name{selectCells}
\alias{selectCells}
\alias{selectCells,DepMapCopyNumber-method}
\alias{selectCells,DepMapExpression-method}
\alias{selectCells,DepMapGeneEffect-method}
\title{Select cells}
\usage{
selectCells(object, ...)

\S4method{selectCells}{DepMapCopyNumber}(object, ...)

\S4method{selectCells}{DepMapExpression}(object, ...)

\S4method{selectCells}{DepMapGeneEffect}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Key value pairs that map to cell line metadata defined in \code{colData}.}
}
\value{
Modified object.
}
\description{
Select cells
}
\note{
Updated 2023-09-12.
}
\examples{
data(crispr)

## DepMapGeneEffect ====
object <- crispr
dim(object)
lineage <- colData(object)[["lineage"]][[1L]]
print(lineage)
all(colData(object)[["lineage"]] == lineage)
object <- selectCells(object, lineage = lineage)
all(colData(object)[["lineage"]] == lineage)
dim(object)
}
\seealso{
\itemize{
\item \code{AcidExperiment::selectSamples}.
}
}
