% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show}
\alias{show}
\alias{show,DepMapFusion-method}
\alias{show,DepMapGeneEffect-method}
\alias{show,DepMapMutations-method}
\title{Show an object}
\usage{
\S4method{show}{DepMapFusion}(object)

\S4method{show}{DepMapGeneEffect}(object)

\S4method{show}{DepMapMutations}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Console output.
}
\description{
Show an object
}
\note{
Updated 2023-08-02.
}
\examples{
data(crispr)

## GeneEffect ====
show(crispr)
}
