% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show}
\alias{show}
\alias{show,CCLEFusionData-method}
\alias{show,CCLEMutationData-method}
\alias{show,GeneEffect-method}
\title{Show an object}
\usage{
\S4method{show}{CCLEFusionData}(object)

\S4method{show}{CCLEMutationData}(object)

\S4method{show}{GeneEffect}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Console output.
}
\description{
Show an object
}
\note{
Updated 2022-09-21.
}
\examples{
data(crispr)

## GeneEffect ====
show(crispr)
}
